+++ sdk/src/wallet/account/operations/transaction/prepare_transaction.rs
@@ -25,9 +25,10 @@ impl AccountHandle {
     pub async fn prepare_transaction(
         &self,
         outputs: Vec<Output>,
-        options: Option<TransactionOptions>,
+        options: impl Into<Option<TransactionOptions>>,
     ) -> crate::wallet::Result<PreparedTransactionData> {
         log::debug!("[TRANSACTION] prepare_transaction");
+        let options = options.into();
         let prepare_transaction_start_time = Instant::now();
         let rent_structure = self.client.get_rent_structure().await?;
         let token_supply = self.client.get_token_supply().await?;