+++ sdk/src/wallet/bindings/nodejs/examples/4c-send-micro-transaction.js
@@ -21,6 +21,9 @@ async function run() {
                 address,
                 amount,
             },
+            {
+                allowMicroAmount: true,
+            }
         ]);
 
         console.log(response);