+++ sdk/src/wallet/bindings/nodejs/examples/7-events.js
@@ -29,6 +29,9 @@ async function run() {
                     'rms1qph3f6y3ps7zccucatf70y37kz7udzp94aefg6mzxdgpa5xxerg9u4s0xyz',
                 amount: '1000',
             },
+            {
+                allowMicroAmount: true,
+            }
         ]);
 
         // provide event type to remove only event listeners of this type