+++ sdk/src/wallet/bindings/nodejs/types/transactionOptions.ts
@@ -1,4 +1,4 @@
-import type { ITaggedDataPayload, HexEncodedAmount } from '@iota/types';
+import type { ITaggedDataPayload, HexEncodedAmount, Burn } from '@iota/types';
 
 /** Options for the transaction creation */
 export interface TransactionOptions {
@@ -6,8 +6,11 @@ export interface TransactionOptions {
     taggedDataPayload?: ITaggedDataPayload;
     /** Custom inputs that should be used for the transaction */
     customInputs?: string[];
+    mandatoryInputs?: string[];
+    burn?: Burn;
     /** Optional note, that is only stored locally */
     note?: string;
+    allowMicroAmount: boolean;
 }
 
 /** The RemainderValueStrategy */
@@ -35,7 +38,7 @@ export type CustomAddress = {
 };
 
 /** Native token options for minting more native tokens */
-export interface IncreaseNativeTokenSupplyOptions {}
+export interface IncreaseNativeTokenSupplyOptions { }
 
 /** Native token options for minting */
 export interface NativeTokenOptions {